+++ packages/styles/projects/emotion/src/typography/proseHeading3XlStyle.ts
@@ -1,5 +1,7 @@
-import { fontPorscheNext, fontWeightNormal, leadingNormal, typescale3Xl } from '@porsche-design-system/tokens';
+import { colorPrimary } from '../color';
+import { fontPorscheNext, fontWeightNormal, leadingNormal, typescale3Xl } from '../font';
 
 export const proseHeading3XlStyle = {
-  font: `normal normal ${fontWeightNormal} ${typescale3Xl} / ${leadingNormal} ${fontPorscheNext}`,
+  font: `${fontWeightNormal} ${typescale3Xl} / ${leadingNormal} ${fontPorscheNext}`,
+  color: colorPrimary,
 };