+++ packages/styles/projects/vanilla-extract/src/typography/proseHeadingSmStyle.ts
@@ -1,5 +1,7 @@
-import { fontPorscheNext, fontWeightSemibold, leadingNormal, typescaleSm } from '@porsche-design-system/tokens';
+import { colorPrimary } from '../color';
+import { fontPorscheNext, fontWeightSemibold, leadingNormal, typescaleSm } from '../font';
 
 export const proseHeadingSmStyle = {
-  font: `normal normal ${fontWeightSemibold} ${typescaleSm} / ${leadingNormal} ${fontPorscheNext}`,
+  font: `${fontWeightSemibold} ${typescaleSm} / ${leadingNormal} ${fontPorscheNext}`,
+  color: colorPrimary,
 };