+++ packages/tokens/src/spacing/static/spacingStatic2Xs.ts
@@ -1,0 +1,1 @@
+export const spacingStatic2Xs = '1px';